pyproject.toml
[project]
name = "transcript-trainer-backend"
version = "0.1.0"
description = "Backend for Transcript Trainer"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.117.1",
]

[dependency-groups]
dev = [
    "bandit>=1.8.6",
    "black>=25.9.0",
    "flake8>=7.3.0",
    "isort>=6.0.1",
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
    "pytest-asyncio>=1.2.0",
    "pytest-vcr>=1.0.2",
    "youtube-transcript-api>=1.2.2",
]
